Added create post form and handles its functionallity
