+++ app/main/forms.py
@@ -21,12 +21,6 @@ class EditProfileForm(FlaskForm):
                 raise ValidationError('Please user a different username.')
 
 
-class PostForm(FlaskForm):
-    title = StringField('Say something', validators=[DataRequired(), Length(min=1, max=140)])
-    post = TextAreaField('Say something', validators=[DataRequired()])
-    submit = SubmitField('Submit')
-
-
 class SearchForm(FlaskForm):
     q = StringField('Search', validators=[DataRequired()])
