+++ app/main/routes.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from app import db
 from app.models import User, Post
-from app.main.forms import EditProfileForm, PostForm, SearchForm
+from app.main.forms import EditProfileForm, SearchForm
 from flask import render_template, flash, redirect, url_for, request, current_app, g
 from flask_login import current_user, login_required
 from app.main import bp
@@ -19,13 +19,6 @@ def before_request():
 @bp.route('/index', methods=['GET', 'POST'])
 @login_required
 def index():
-    form = PostForm()
-    if form.validate_on_submit():
-        post = Post(title=form.title.data,body=form.post.data, author=current_user)
-        db.session.add(post)
-        db.session.commit()
-        flash('Your post is now live!')
-        return redirect(url_for('main.index'))
     page = request.args.get('page', 1, type=int)
     posts = current_user.followed_posts()\
         .paginate(page, current_app.config['POSTS_PER_PAGE'], False)
@@ -35,7 +28,6 @@ def index():
         if posts.has_prev else None
     return render_template('index.html',
                            title='Home',
-                           form=form,
                            posts=posts.items,
                            next_url=next_url,
                            prev_url=prev_url)
