+++ app/post/forms.py
@@ -1,0 +1,9 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField, TextAreaField
+from wtforms.validators import DataRequired, Length
+
+
+class PostForm(FlaskForm):
+    title = StringField('Title', validators=[DataRequired(), Length(min=1, max=140)])
+    post = TextAreaField('Content', validators=[DataRequired()])
+    submit = SubmitField('Post')
