+++ app/templates/index.html
@@ -1,15 +1,6 @@
 {% extends "base.html" %}
-{% import 'bootstrap/wtf.html' as wtf %}
-
 {% block app_content %}
     <h1>Hi, {{ current_user.username }}!</h1>
-    {% if form %}
-    <div class="row">
-        <div class="col-md-4">
-            {{ wtf.quick_form(form) }}
-        </div>
-    </div>
-    {% endif %}
     {% for post in posts %}
         {% include '_includes/post.html' %}
     {% endfor %}
